fsp: Fix msg vaargs usage
hw/fsp/fsp.c:1011:17: warning: passing an object that undergoes default argument promotion to
      'va_start' has undefined behavior [-Wvarargs]
        va_start(list, add_words);
                       ^
hw/fsp/fsp.c:1007:59: note: parameter of type 'u8' (aka 'unsigned char') is declared here
void fsp_fillmsg(struct fsp_msg *msg, u32 cmd_sub_mod, u8 add_words, ...)
                                                          ^
	[CC]  platforms/ibm-fsp/apollo-pci.o
hw/fsp/fsp.c:1026:17: warning: passing an object that undergoes default argument promotion to
      'va_start' has undefined behavior [-Wvarargs]
        va_start(list, add_words);
                       ^
hw/fsp/fsp.c:1016:47: note: parameter of type 'u8' (aka 'unsigned char') is declared here
struct fsp_msg *fsp_mkmsg(u32 cmd_sub_mod, u8 add_words, ...)

+++ hw/fsp/fsp.c
@@ -1004,7 +1004,7 @@ static void __fsp_fillmsg(struct fsp_msg *msg, u32 cmd_sub_mod,
 	va_end(list);
 }
 
-void fsp_fillmsg(struct fsp_msg *msg, u32 cmd_sub_mod, u8 add_words, ...)
+void fsp_fillmsg(struct fsp_msg *msg, u32 cmd_sub_mod, u32 add_words, ...)
 {
 	va_list list;
 
@@ -1013,7 +1013,7 @@ void fsp_fillmsg(struct fsp_msg *msg, u32 cmd_sub_mod, u8 add_words, ...)
 	va_end(list);
 }
 
-struct fsp_msg *fsp_mkmsg(u32 cmd_sub_mod, u8 add_words, ...)
+struct fsp_msg *fsp_mkmsg(u32 cmd_sub_mod, u32 add_words, ...)
 {
 	struct fsp_msg *msg = fsp_allocmsg(!!(cmd_sub_mod & 0x1000000));
 	va_list list;
